Split AdminService into domain-specific services
Replace monolithic AdminService with four focused services following
the Single Responsibility Principle:
- AdminEventService: event CRUD, ticket generation, Spotify/TM operations
- AdminUserService: user listing, role updates, status management
- AdminOrderService: order listing, revenue calculation, recent orders
- AdminDashboardService: aggregates stats from the other three services

Update AdminController to inject domain-specific services. Update all
tests (controller, sync, authorization bypass) to mock the new services.
