+++ backend/src/main/java/com/mockhub/admin/service/AdminUserService.java
@@ -1,0 +1,102 @@
+package com.mockhub.admin.service;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.mockhub.auth.dto.UserDto;
+import com.mockhub.auth.entity.Role;
+import com.mockhub.auth.entity.User;
+import com.mockhub.auth.repository.RoleRepository;
+import com.mockhub.auth.repository.UserRepository;
+import com.mockhub.common.dto.PagedResponse;
+import com.mockhub.common.exception.ResourceNotFoundException;
+
+@Service
+public class AdminUserService {
+
+    private static final Logger log = LoggerFactory.getLogger(AdminUserService.class);
+
+    private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
+
+    public AdminUserService(UserRepository userRepository, RoleRepository roleRepository) {
+        this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
+    }
+
+    @Transactional(readOnly = true)
+    public PagedResponse<UserDto> getAllUsers(Pageable pageable) {
+        Page<User> userPage = userRepository.findAll(pageable);
+
+        List<UserDto> content = userPage.getContent().stream()
+                .map(this::toUserDto)
+                .toList();
+
+        return new PagedResponse<>(
+                content,
+                userPage.getNumber(),
+                userPage.getSize(),
+                userPage.getTotalElements(),
+                userPage.getTotalPages()
+        );
+    }
+
+    @Transactional(readOnly = true)
+    public long getUserCount() {
+        return userRepository.count();
+    }
+
+    @Transactional
+    public void updateUserRoles(Long userId, Set<String> roleNames) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ResourceNotFoundException("User", "id", userId));
+
+        Set<Role> roles = new HashSet<>();
+        for (String roleName : roleNames) {
+            Role role = roleRepository.findByName(roleName)
+                    .orElseThrow(() -> new ResourceNotFoundException("Role", "name", roleName));
+            roles.add(role);
+        }
+
+        user.setRoles(roles);
+        userRepository.save(user);
+        log.atInfo().setMessage("Admin updated roles for user {} to {}").addArgument(userId).addArgument(() -> roleNames.toString().replaceAll("[\\r\\n]", "")).log();
+    }
+
+    @Transactional
+    public void updateUserStatus(Long userId, boolean enabled) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ResourceNotFoundException("User", "id", userId));
+
+        user.setEnabled(enabled);
+        userRepository.save(user);
+        log.info("Admin {} user {}", enabled ? "enabled" : "disabled", userId);
+    }
+
+    private UserDto toUserDto(User user) {
+        Set<String> roleNames = new HashSet<>();
+        for (Role role : user.getRoles()) {
+            roleNames.add(role.getName());
+        }
+
+        return new UserDto(
+                user.getId(),
+                user.getEmail(),
+                user.getFirstName(),
+                user.getLastName(),
+                user.getPhone(),
+                user.getAvatarUrl(),
+                user.isEmailVerified(),
+                roleNames,
+                user.getCreatedAt()
+        );
+    }
+}
